Generalize default constructors for TypeTensor and TypeVector
Assigning to the (template typed) values in these data structures with
the (integer) zero doesn't work when the template type is not POD nor
has an overloaded assignment operator taking an int/number.  Assigning
"{}" instead works for a larger set of types, while still giving zero
values to POD.  This issue was run into when I tried to use an
eigen::Matrix<DualNumber<...>> in TypeVector.