Fix SFC Partitioner on distributed meshes
I ran across this speedbump while trying out some alternate configs,
which I needed for debugging the quad-precision support additions I made
recently, which I did to help debugging the newest IGA code.

The problem has now been whipped.  We often have holes in our id
numbering in the distributed mesh case, so to index by id we need
vectors that can hold the maximum id number of an element, not just the
number of elements.

Anyone who's reading this commit message has decided to inspect not flee
the tale of the flog of the bump in our logs on the hole in the bottom
of the C.